Add files via upload
Gauss Elimination method is a method of solving matrix form of set of algebraic equations whole solutions are unique. this is a direct method where elimination of variables or reduction of equations is obtained step by step.
python code is provided here to help us to solve the simultaneous equations.